git-svn-id: svn://srv-dc/svn2/Citrus@185012 20bf7d6a-bd21-9d4d-a3c4-a2c14bb8c68a

+++ Lime/Source/NumericExtensions.cs
@@ -32,7 +32,7 @@ namespace Lime
 			return Mathf.Sqrt(x);
 		}
 
-		public static float Sign(this float x)
+		public static int Sign(this float x)
 		{
 			return Math.Sign(x);
 		}
